+++ instagram_to_discord/const_value.py
@@ -1,0 +1,1 @@
+FSIZE_TARGET = 2 ** 23 - 100
